file-vali/.ipynb_checkpoints/requirements-checkpoint.txt
flask
flask-cors
pandas
google-cloud-storage
werkzeug
